:tada: :wrench: The menu in window dynamically changes syntax highlighting
Also, refactored `theme_manager.py` to split it into `theme.py`.
